requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5.11
langchain==0.3.3
langchain_chroma==0.1.4
langchain_community==0.3.2
langchain_core==0.3.10
langchain_experimental==0.3.2
langchain_groq==0.2.0
langchain_nvidia_ai_endpoints==0.3.0
langchain_openai==0.2.2
langchain_text_splitters==0.3.0
langgraph==0.2.35
llama_index==0.11.17
Markdown==3.7
openai==1.51.2
pandas==1.5.3
Pillow==10.4.0
plotly==5.24.1
Requests==2.32.3
sentence_transformers==3.1.1
SQLAlchemy==2.0.31
SQLAlchemy==2.0.35
sqldf==0.4.2
streamlit==1.38.0
streamlit==1.39.0
